+++ test/Virtlink.Utilib.Tests/Collections/ArraysTests.cs
@@ -1,9 +1,0 @@
-﻿namespace Virtlink.Utilib.Collections
-{
-    /// <summary>
-    /// Tests the <see cref="Arrays"/> class.
-    /// </summary>
-    public partial class ArraysTests
-    {
-    }
-}